+++ gdqn/envs/__init__.py
@@ -1,0 +1,3 @@
+from gymnasium.envs.registration import register
+
+register(id='trading-v0', entry_point='env:TradingEnv')
