--- utils/envs/__init__.py
@@ -1,3 +1,0 @@
-from gymnasium.envs.registration import register
-
-register(id='trading-v0', entry_point='utils.envs.env:TradingEnv')
